cli docs: preliminary.
These code blocks are generated by testmark rigging in the warpforge
implementation repo, and contain fairly literally the actual CLI help.
This content then gets hoisted back out of codeblock form during the
rendering pipeline of the static site tooling, and rendered as
regular markdown.  This lets us autoupdate and have nice presentation
at the same time.

This actual content is somewhat preliminary.  We're still going through
R&D of how exactly the warpforge repo will emit this stuff.
The general outline is probably about right (use of headings, etc);
things like whitespace are absolutely not finalized.
Crosslinking to subcommands is entirely not yet present, but desired.
(I'm still trying to figure out how to sneak linking through the other
half of this -- the markdown-to-terminal-ansi gadgets which we're
rolling out in the warpforge implementation.)

The css here will need a lot of work.  None is applied so far.
I think I'd probably like it to look as similar to CLI helptext as
possible... just without the foibels of fixed wrapping.