drop generated CLI content pages.  Need to reconsider lifecycle of these.
The original idea was to have these be generated by testmark-powered
tests that live in the warpforge implementation repo...
which is a good idea, and maybe we'll someday return to it, BUT...

It's also overridingly important that the tests in that repo stand
alone in that repo.  So the fixtures need to live there first.

If we then replicate that data into this repo, that's fine and dandy.
But having it live here first, and create an unstable relationship
that crosses the two repos without good versioning... No.
And unfortunately that was the implementation that caused these
files to exist.

We don't have scripts to keep these files up to date yet.
So they go.
Maybe we write such scripts in the future, but today they do not exist.